refactor to extract_schema
