+++ sycamore/docset.py
@@ -8,7 +8,7 @@ from sycamore.data import Document
 from sycamore.plan_nodes import Node
 from sycamore.transforms.embed import Embedder
 from sycamore.transforms.extract_entity import EntityExtractor
-from sycamore.transforms.schema import SchemaExtractor, PropertyExtractor
+from sycamore.transforms.extract_schema import SchemaExtractor, PropertyExtractor
 from sycamore.transforms.partition import Partitioner
 from sycamore.transforms.summarize import Summarizer
 from sycamore.transforms.extract_table import TableExtractor
