+++ sycamore/transforms/__init__.py
@@ -8,7 +8,7 @@ from sycamore.transforms.extract_table import TableExtractor
 from sycamore.transforms.spread_properties import SpreadProperties
 from sycamore.transforms.summarize import Summarize
 from sycamore.transforms.merge_elements import Merge
-from sycamore.transforms.schema import (
+from sycamore.transforms.extract_schema import (
     ExtractSchema,
     ExtractBatchSchema,
     SchemaExtractor,
